Dropped association from Customer to SageImport and SageDownload
Reasoning is people want the Id field for a Customer in Sage to be
in a format of their choosing, unless this is stored we then have
no way of mapping back at the Import stage to create the
association.

Easier is to not associate and just look for Customers associated
with the Order we're processing, then they are flagged automatically
when we associate and order with an Import into sage.